+++ src/fhir/search-helpers.ts
@@ -160,8 +160,10 @@ function patientSearchParams(
   switch (resourceType) {
     case "Coverage":
       return { "beneficiary.identifier": identifier };
+    case "AllergyIntolerance":
     case "Condition":
     case "Encounter":
+    case "Immunization":
     case "MedicationAdministration":
     case "MedicationDispense":
     case "MedicationRequest":