+++ src/index.ts
@@ -1,7 +1,10 @@
 // Content components
+export * from "@/components/content/allergies/patient-allergies";
+export * from "@/components/content/allergies/patient-allergies-column";
 export * from "@/components/content/conditions";
 export * from "@/components/content/conditions/patient-conditions";
 export * from "@/components/content/conditions-table-base";
+export * from "@/components/content/immunizations/patient-immunizations";
 export * from "@/components/content/medication-drawer";
 export * from "@/components/content/medications-table-base";
 export * from "@/components/content/medications/add-new-med-drawer";